backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
pydantic[email]
httpx
requests
langgraph
langchain
langchain-core
python-dotenv
websockets
serpapi
openai
tavily-python
passlib
python-jose[cryptography]
python-multipart
google-search-results
langgraph-checkpoint-sqlite
bs4
duckduckgo-search
graphviz
pygraphviz
networkx
langchain-anthropic
langchain-community
langchain-openai
bcrypt
transformers
sentencepiece
torch
yake
bs4
duckduckgo
ddgs
google-search-results>=2.4.2